// src/server/api/models/insurance/models/Quote.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { InsuranceProductType } from './InsuranceProductType';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { PolicyPaymentTerms } from './PolicyPaymentTerms';
import type { PolicyRiskType } from './PolicyRiskType';
import type { QuoteStatus } from './QuoteStatus';
import type { UUID } from './UUID';

export type Quote = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    /**
     * A unique quote number
     */
    quoteNo: string;
    /**
     * A quote number generated by the insurance partner
     */
    refQuoteNo: string;
    /**
     * The quote description
     */
    quoteLabel: string;
    /**
     * The quote currency
     */
    currency: string;
    /**
     * Notes related to the record
     */
    notes?: string | null;
    /**
     * The status of the quote
     */
    status: QuoteStatus;
    /**
     * The type of risk insured
     */
    riskType: PolicyRiskType;
    /**
     * Indicates the payment terms for the premium
     */
    paymentTerms: PolicyPaymentTerms;
    /**
     * The premium charged for the quote
     */
    premium: number | null;
    /**
     * The commission charged for the quote
     */
    commission: number | null;
    /**
     * The tax charged for the quote
     */
    tax: number | null;
    /**
     * The total amount charged for the quote
     */
    total: number | null;
    /**
     * The value insured
     */
    sumInsured: number | null;
    /**
     * The amount due in each payment period
     */
    scheduledAmount: number | null;
    /**
     * Indicates the start date for the quote
     */
    startDate: LocalDate;
    /**
     * Indicates the end date for the quote
     */
    endDate: LocalDate;
    /**
     * The quote product ID
     */
    productId: UUID | null;
    /**
     * The quote product label
     */
    productLabel: string | null;
    /**
     * The type of the quote
     */
    productType: InsuranceProductType;
    branchId: UUID;
    branchLabel: string;
    partnerId?: UUID | null;
    partnerLabel?: string | null;
};